Add examples folder with sample data generation and executed notebooks
- Created comprehensive sample data generator using Faker
- Generated sample data for all 6 data types (HRIS, email, Slack, calendar, documents, code)
- Created 4 example notebooks with executed outputs:
  * 01_basic_analysis.ipynb - Getting started guide
  * 02_network_metrics.ipynb - Network metrics deep dive
  * 03_community_detection.ipynb - Community detection examples
  * 04_visualization.ipynb - Network visualization examples
- Fixed data ingestion to handle NaN values properly
- Fixed data processors to handle mixed types in sorting
- Updated .gitignore to track example notebooks (removed *.ipynb exclusion)
- Added faker and jupyter to requirements.txt
- Created comprehensive README and SETUP guides for examples

+++ orgnet/graph/weights.py
@@ -114,7 +114,7 @@ class EdgeWeightCalculator:
             DataFrame with weighted meeting matrix
         """
         coattendance = self.processor.compute_meeting_coattendance(meetings)
-        weighted = coattendance.copy()
+        weighted = coattendance.copy().astype(float)
 
         # Weight by meeting size and duration
         for meeting in meetings:
